butiful planet

+++ Project/OPENGL-Course/Real OpenGL Project/Window.h
@@ -14,8 +14,8 @@ public:
 	~Window();
 
 	int Initialise();
-	GLfloat getBufferWidth() { return bufferWidth; }
-	GLfloat getBufferHeight() { return bufferHeight; }
+	GLfloat getBufferWidth() { return (GLfloat)bufferWidth; }
+	GLfloat getBufferHeight() { return (GLfloat)bufferHeight; }
 	bool getShouldClose() { return glfwWindowShouldClose(mainWindow); }
 	GLFWwindow* getWindow() { return mainWindow; }
